backend/requirements.txt
python-multipart
fastapi==0.109.0
uvicorn==0.27.0.post1
langchain==0.1.19
langgraph==0.0.20
langsmith==0.1.0
neo4j==5.17.0
python-dotenv==1.0.1
requests==2.31.0
langchain-upstage==0.1.0
langchain_core==0.1.52
langchain_community==0.0.38
beautifulsoup4==4.12.3
